Code review updates round 2.
