+++ spec/ddtrace/contrib/httparty/patcher_spec.rb
@@ -5,16 +5,16 @@ require 'httparty/request'
 
 RSpec.describe Datadog::Contrib::HTTParty::Patcher do
   describe '.patch' do
-    it 'adds RequestPatch to ancestors of Request class' do
+    it 'adds Instrumentation::Request to ancestors of Request class' do
       described_class.patch
 
-      expect(HTTParty::Request.ancestors).to include(Datadog::Contrib::HTTParty::RequestPatch)
+      expect(HTTParty::Request.ancestors).to include(Datadog::Contrib::HTTParty::Instrumentation::Request)
     end
 
     it 'adds ModulePatch to ancestors of HTTParty::ClassMethods' do
       described_class.patch
 
-      expect(HTTParty::ClassMethods.ancestors).to include(Datadog::Contrib::HTTParty::ModulePatch)
+      expect(HTTParty::ClassMethods.ancestors).to include(Datadog::Contrib::HTTParty::Instrumentation::Helpers)
     end
   end
 end
